Fixed some bugs into multicurves. Added one dimensional curves (same types of 2D curves) Added initial support for animations; we started introducing GAnimElement and GProperty classes. Then some basic scalar animated properties have been implemented (hermite/linear/constant interpolators). Added a simple example (applied to colors) to show animation capabilities.

+++ trunk/vs2003/examples/opengl/bezier2d/drawbezier.cpp
@@ -491,5 +491,5 @@ int WINAPI WinMain(	HINSTANCE	hInstance,			// Instance
 	// Shutdown
 	KillGLWindow();									// Kill The Window
 	KillApp();
-	return (msg.wParam);							// Exit The Program
+	return (int)(msg.wParam);							// Exit The Program
 }
